requirements.txt
# NLP & Machine Learning
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
sentencepiece==0.1.99

# Web Scraping
beautifulsoup4==4.12.0
requests==2.31.0
selenium==4.15.0
trafilatura==1.7.0
lxml==4.9.3

# Data Processing
pandas==2.1.0
numpy==1.26.0
nltk==3.8.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Scheduling
apscheduler==3.10.4

# Dashboard
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
tenacity==8.2.3

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
